// src/api/generated/model/hydrationDailyResponseStatus.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Fitness API
 * REST API for the Fitness & Nutrition SPA. Call GET /sanctum/csrf-cookie then POST /api/auth/login before accessing protected endpoints.
 * OpenAPI spec version: 1.0.0
 */

export type HydrationDailyResponseStatus = typeof HydrationDailyResponseStatus[keyof typeof HydrationDailyResponseStatus];


export const HydrationDailyResponseStatus = {
  no_data: 'no_data',
  unknown: 'unknown',
  partial: 'partial',
  complete: 'complete',
} as const;